without properties now that we're authenticating with keypair
