+++ readthedocs/code-tabs/java/src/main/java/nl/esciencecenter/xenon/tutorial/SlurmQueuesGetter.java
@@ -1,7 +1,5 @@
 package nl.esciencecenter.xenon.tutorial;
 
-import java.util.Map;
-
 import nl.esciencecenter.xenon.XenonException;
 import nl.esciencecenter.xenon.credentials.Credential;
 import nl.esciencecenter.xenon.credentials.PasswordCredential;
@@ -14,12 +12,10 @@ public class SlurmQueuesGetter {
 
         String host = "localhost";
         String port = "10022";
-        Map<String, String> properties = null;
-
-        runExample(host, port, properties);
+        runExample(host, port);
     }
 
-    public static void runExample(String host, String port, Map<String, String> properties) throws XenonException {
+    public static void runExample(String host, String port) throws XenonException {
 
         String adaptor = "slurm";
         String location = "ssh://" + host + ":" + port;
@@ -30,7 +26,7 @@ public class SlurmQueuesGetter {
         Credential credential = new PasswordCredential(username, password);
 
         // create the SLURM scheduler
-        try (Scheduler scheduler = Scheduler.create(adaptor, location, credential, properties)) {
+        try (Scheduler scheduler = Scheduler.create(adaptor, location, credential)) {
             // ask SLURM for its queues
             System.out.println("Available queues (starred queue is default):");
             for (String queueName : scheduler.getQueueNames()) {
