+++ readthedocs/code-tabs/java/src/main/java/nl/esciencecenter/xenon/tutorial/UploadFileLocalToSftpAbsolutePathsWebdavEquivalent.java
@@ -1,7 +1,5 @@
 package nl.esciencecenter.xenon.tutorial;
 
-import java.util.Map;
-
 import nl.esciencecenter.xenon.XenonException;
 import nl.esciencecenter.xenon.credentials.PasswordCredential;
 import nl.esciencecenter.xenon.filesystems.CopyMode;
@@ -16,12 +14,10 @@ public class UploadFileLocalToSftpAbsolutePathsWebdavEquivalent {
 
         String host = "localhost";
         String port = "10080";
-        Map<String, String> properties = null;
-
-        runExample(host, port, properties);
+        runExample(host, port);
     }
 
-    public static void runExample(String host, String port, Map<String, String> properties) throws XenonException {
+    public static void runExample(String host, String port) throws XenonException {
 
         // use the local file system adaptor to create a file system representation
         String adaptorLocal = "file";
@@ -39,7 +35,7 @@ public class UploadFileLocalToSftpAbsolutePathsWebdavEquivalent {
         String adaptorRemote = "webdav";
         String location = "http://" + host + ":" + port;
 
-        FileSystem filesystemRemote = FileSystem.create(adaptorRemote, location, credential, properties);
+        FileSystem filesystemRemote = FileSystem.create(adaptorRemote, location, credential);
 
         // define which file to upload to
         Path fileRemote = new Path("/~xenon/uploads/sleep.sh");
