+++ readthedocs/code-tabs/java/src/test/java/nl/esciencecenter/xenon/tutorial/AllTogetherNowWrongTest.java
@@ -1,42 +1,27 @@
 package nl.esciencecenter.xenon.tutorial;
 
-import java.util.HashMap;
-import java.util.Map;
-
-import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
-import org.testcontainers.containers.GenericContainer;
+import org.testcontainers.containers.FixedHostPortGenericContainer;
+import org.testcontainers.containers.wait.strategy.Wait;
 
 
 public class AllTogetherNowWrongTest {
 
-    private String host;
-    private String port;
-    private Map<String, String> propertiesSsh;
-    private Map<String, String> propertiesSftp;
+    private final String image = "xenonmiddleware/slurm:17";
+    private final String host = "localhost";
+    private final String port = "10022";
 
     @Rule
-    public GenericContainer<?> slurm = new GenericContainer<>("xenonmiddleware/slurm:17").withExposedPorts(22);
+    public FixedHostPortGenericContainer<?> slurm = new FixedHostPortGenericContainer<>(image)
+                                                        .withFixedExposedPort(Integer.parseInt(port), 22)
+                                                        .waitingFor(Wait.forHealthcheck());
 
-    @Before
-    public void setUp() {
-        host = slurm.getContainerIpAddress();
-        port = Integer.toString(slurm.getFirstMappedPort());
-
-        propertiesSsh = new HashMap<String, String>();
-        propertiesSsh.put("xenon.adaptors.schedulers.ssh.loadSshConfig", "false");
-        propertiesSsh.put("xenon.adaptors.schedulers.ssh.loadKnownHosts", "false");
-
-        propertiesSftp = new HashMap<String, String>();
-        propertiesSftp.put("xenon.adaptors.filesystems.sftp.loadSshConfig", "false");
-        propertiesSftp.put("xenon.adaptors.filesystems.sftp.loadKnownHosts", "false");
-
-    }
 
     @Test(expected = Test.None.class)
     public void test1() throws Exception {
-        AllTogetherNowWrong.runExample(host, port, propertiesSsh, propertiesSftp);
+        AllTogetherNowWrong.runExample(host, port);
     }
+
 
 }
