+++ readthedocs/code-tabs/java/src/test/java/nl/esciencecenter/xenon/tutorial/DownloadFileSftpToLocalAbsolutePathsWebdavEquivalentTest.java
@@ -1,36 +1,33 @@
 package nl.esciencecenter.xenon.tutorial;
 
 import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
 
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
-import org.testcontainers.containers.GenericContainer;
+import org.testcontainers.containers.FixedHostPortGenericContainer;
+import org.testcontainers.containers.wait.strategy.Wait;
 
 
 public class DownloadFileSftpToLocalAbsolutePathsWebdavEquivalentTest {
 
-    private String host;
-    private String port;
-    private Map<String, String> properties;
+    private final String image = "xenonmiddleware/webdav";
+    private final String host = "localhost";
+    private final String port = "10080";
 
     @Rule
-    public GenericContainer<?> webdav = new GenericContainer<>("xenonmiddleware/webdav").withExposedPorts(80);
+    public FixedHostPortGenericContainer<?> webdav = new FixedHostPortGenericContainer<>(image)
+                                                        .withFixedExposedPort(Integer.parseInt(port), 80)
+                                                        .waitingFor(Wait.forHealthcheck());
 
     @Before
     public void setUp() throws UnsupportedOperationException, IOException, InterruptedException {
-        host = webdav.getContainerIpAddress();
-        port = Integer.toString(webdav.getFirstMappedPort());
-        properties = new HashMap<String, String>();
-
         // test tries to download this file, so we need to make it first
         webdav.execInContainer("touch", "/home/xenon/sleep.stdout.txt");
     }
 
     @Test(expected = Test.None.class)
     public void test1() throws Exception {
-        DownloadFileSftpToLocalAbsolutePathsWebdavEquivalent.runExample(host, port, properties);
+        DownloadFileSftpToLocalAbsolutePathsWebdavEquivalent.runExample(host, port);
     }
 }
